Move LTV to total claim cell in ActiveOffersTable Recalc ltv
