+++ src/pages/OffersPage/components/ActiveOffersTable/TableCells/InterestCell.tsx
@@ -3,26 +3,56 @@ import { FC } from 'react'
 import { calculateCurrentInterestSolPure } from 'fbonds-core/lib/fbond-protocol/functions/perpetual'
 import moment from 'moment'
 
-import { createSolValueJSX } from '@banx/components/TableComponents'
+import { createPercentValueJSX, createSolValueJSX } from '@banx/components/TableComponents'
 
 import { Loan } from '@banx/api/core'
-import { formatDecimal } from '@banx/utils'
+import { HealthColorIncreasing, formatDecimal, getColorByPercent } from '@banx/utils'
+
+import styles from '../ActiveOffersTable.module.less'
 
 interface InterestCellProps {
   loan: Loan
+  isCardView: boolean
 }
 
-export const InterestCell: FC<InterestCellProps> = ({ loan }) => {
+export const InterestCell: FC<InterestCellProps> = ({ loan, isCardView }) => {
   const { solAmount, feeAmount, amountOfBonds, soldAt } = loan.bondTradeTransaction || {}
 
-  const loanValueWithFee = solAmount + feeAmount
+  const totalLoanValue = solAmount + feeAmount
+  const collectionFloor = loan.nft.collectionFloor
 
-  const calculatedInterest = calculateCurrentInterestSolPure({
-    loanValue: loanValueWithFee,
+  const interestParameters = {
+    loanValue: totalLoanValue,
     startTime: soldAt,
     currentTime: moment().unix(),
     rateBasePoints: amountOfBonds,
-  })
+  }
 
-  return createSolValueJSX(calculatedInterest + loanValueWithFee, 1e9, '--', formatDecimal)
+  const currentInterest = calculateCurrentInterestSolPure(interestParameters)
+  const totalClaimValue = currentInterest + totalLoanValue
+
+  const formattedClaimValue = createSolValueJSX(totalClaimValue, 1e9, '--', formatDecimal)
+
+  const loanToValueRatio = (totalClaimValue / collectionFloor) * 100
+
+  return !isCardView ? (
+    <div className={styles.lentInfo}>
+      <span>{formattedClaimValue}</span>
+      {createLtvValueJSX(loanToValueRatio)}
+    </div>
+  ) : (
+    <span>
+      {formattedClaimValue} ({createLtvValueJSX(loanToValueRatio)})
+    </span>
+  )
+}
+
+const createLtvValueJSX = (ltv: number) => {
+  const colorLTV = getColorByPercent(ltv, HealthColorIncreasing)
+
+  return (
+    <span className={styles.lentInfoSubtitle} style={{ color: colorLTV }}>
+      {createPercentValueJSX(ltv)} LTV
+    </span>
+  )
 }
